tazebao/requirements/local.txt
# Local development dependencies go here
-r common.txt

django-debug-toolbar>=1.5
sqlparse>=0.2.1
